chore(user): extract notification settings validation
Generated-By: PostHog Desktop
Task-Id: 88465e13-ac11-4b1f-b37c-b074869406ae
